Refactor: collected common functions in embedding_utils.py and removed them from individual embedder scripts.
