+++ scripts/antiberta2_cdr3.py
@@ -26,66 +26,22 @@ Dependencies:
 """
 
 import os
-import csv
-import argparse
-from Bio import SeqIO
 import torch
 from transformers import RoFormerTokenizer, RoFormerModel
 from torch.utils.data import DataLoader, TensorDataset
-
+from .embedding_utils import parse_arguments, fasta_to_dict, load_cdr3, export_embeddings, export_sequence_indices
 # Constants
 MODEL_NAME = "alchemab/antiberta2-cssp"
 BATCH_SIZE = 512  # Adjust based on your GPU's memory
 MAX_LENGTH = 200
 
-
-#Parsing command-line arguments for input and output file paths
-def parse_arguments():
-    """Parse command-line arguments for input and output file paths."""
-    PARSER = argparse.ArgumentParser(description="Input path")
-    PARSER.add_argument("--fasta_path", type=str, required=True,
-                        help="Fasta path + filename.fa")
-    PARSER.add_argument("--output_path", type=str, required=True,
-                        help="Output path + filename.pt \nWill output multiple files if multiple layers are specified with '--layers'. Output file is a single tensor or a list of tensors when --pooling is False.")
-    PARSER.add_argument("--cdr3_path", default = None, type=str,
-                        help="Path to the CDR3 CSV file. Only required when calculating CDR3 sequence embeddings.")
-    PARSER.add_argument("--context", default = 0,type=int,
-                        help="Number of amino acids to include before and after CDR3 sequence")
-    PARSER.add_argument("--layers", type=str, nargs='*', default="-1",
-                        help="Representation layers to extract from the model. Default is the last layer. Example: argument '--layers -1 6' will output the last layer and the sixth layer.")
-    PARSER.add_argument('--pooling', type=lambda x: (str(x).lower() == 'true'), default=True,
-                        help="Whether to pool the embeddings or not. Default is True.")
-    return PARSER.parse_args()
-
-def fasta_to_dict(fasta_file):
-    """Convert FASTA file into a dictionary."""
-    print('Loading and batching input sequences...')
-    seq_dict = {}
-    with open(fasta_file) as f:
-        for record in SeqIO.parse(f, 'fasta'):
-            seq_dict[record.id] = " ".join(str(record.seq)) # AA tokens for hugging face models must be space gapped
-            # print progress
-            if len(seq_dict) % 1000 == 0:
-                print(f'{len(seq_dict)} sequences loaded')
-    return seq_dict
-
-def load_cdr3(cdr3_path):
-    """Load CDR3 sequences and store in a dictionary."""
-    if cdr3_path:
-        with open(cdr3_path) as f:
-            reader = csv.reader(f)
-            cdr3_dict = {rows[0]:rows[1] for rows in reader}
-        return cdr3_dict
-    else:
-        return None
-
-def initialize_model(model_name, batch_size):
+def initialize_model(model_name):
     """Initialize the model, tokenizer, and device."""
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     tokenizer = RoFormerTokenizer.from_pretrained(model_name)
     model = RoFormerModel.from_pretrained(model_name).to(device)
     model.eval()
-    return model, device, tokenizer, batch_size
+    return model, device, tokenizer
 
 def load_layers(model, layers):
     """Check if the specified representation layers are valid."""
@@ -197,29 +153,6 @@ def compute_embeddings(data_loader, model, device,
     torch.cuda.empty_cache()
     return mean_representations, sequence_labels
 
-def export_embeddings(mean_representations, layers, output_path, context=None, pooling=True):
-    """Stack representations of each layer into a single tensor and save to output file."""
-    for layer in layers:
-        if context:
-            output_file_layer = output_path.replace('.pt', f'_context_{context}_layer_{layer}.pt')
-        else:
-            output_file_layer = output_path.replace('.pt', f'_layer_{layer}.pt')
-        if pooling:
-            mean_representations[layer] = torch.vstack(mean_representations[layer])
-        else:
-            output_file_layer = output_file_layer.replace('.pt', '_full.pt')
-        torch.save(mean_representations[layer], output_file_layer)
-        print(f"Saved mean representations for layer {layer} to {output_file_layer}")
-
-def export_sequence_indices(sequence_labels, output_path):
-    """Save sequence indices to a CSV file."""
-    output_file_idx = output_path.replace('.pt', '_idx.csv')
-    with open(output_file_idx, 'w') as f:
-        f.write('index,sequence_id\n')
-        for i, label in enumerate(sequence_labels):
-            f.write(f'{i},{label}\n')
-    print(f"Saved sequence indices to {output_file_idx}")
-
 def main():
     args = parse_arguments()
     # Store arguments
@@ -227,7 +160,7 @@ def main():
     output_path = args.output_path
     cdr3_path = args.cdr3_path
     context = args.context
-    layers = list(map(int, args.layers[0].split()))
+    layers = list(map(int, args.layers))
     pooling = bool(args.pooling)
     # Print summary of arguments
     print(f"FASTA file: {fasta_path}")
@@ -251,19 +184,19 @@ def main():
     cdr3_dict = load_cdr3(cdr3_path)
 
     # Initialize model and prepare input data
-    model, device, tokenizer, batch_size = initialize_model(MODEL_NAME, BATCH_SIZE)
+    model, device, tokenizer = initialize_model(MODEL_NAME)
     layers = load_layers(model, layers)
-    data_loader = load_data(sequences, tokenizer, batch_size, max_length=MAX_LENGTH)
+    data_loader = load_data(sequences, tokenizer, BATCH_SIZE, max_length=MAX_LENGTH)
 
     # Compute embeddings
     mean_representations, sequence_labels = compute_embeddings(
         data_loader, model, device, layers, sequences,
-        context, pooling, cdr3_path, cdr3_dict, batch_size
+        context, pooling, cdr3_path, cdr3_dict, BATCH_SIZE
         )
 
     # Write embeddings to disk and export sequence indices
     export_embeddings(mean_representations, layers, output_path, context, pooling)
-    export_sequence_indices(sequence_labels, output_path)
+    export_sequence_indices(sequence_labels, output_path) 
 
 if __name__ == "__main__":
     main()
