+++ scripts/esm2_cdr3.py
@@ -24,71 +24,20 @@ Dependencies:
     - esm (Facebook Research)
     - torch.utils.data (PyTorch)
 """
+
 import os
 import argparse
-import csv
 import torch
-from Bio import SeqIO
 from esm import FastaBatchedDataset, pretrained
+from .embedding_utils import parse_arguments, fasta_to_dict, load_cdr3, export_embeddings, export_sequence_indices
 
 # Constants
 MODEL_NAME = "esm2_t33_650M_UR50D"
-BATCH_SIZE = 8192 # works with Nvidia V100-32GB GPU
-
-# Parsing command-line arguments for input and output file paths
-def parse_arguments():
-    """Parse command-line arguments for input and output file paths."""
-    PARSER = argparse.ArgumentParser(description="Input path")
-    PARSER.add_argument("--fasta_path", type=str, required=True,
-                        help="Fasta path + filename.fa")
-    PARSER.add_argument("--output_path", type=str, required=True,
-                        help="Output path + filename.pt \nWill output multiple files if multiple layers are specified with '--layers'. Output file is a single tensor or a list of tensors when --pooling is False.")
-    PARSER.add_argument("--cdr3_path", default = None, type=str,
-                        help="Path to the CDR3 CSV file. Only required when calculating CDR3 sequence embeddings.")
-    PARSER.add_argument("--context", default = 0,type=int,
-                        help="Number of amino acids to include before and after CDR3 sequence")
-    PARSER.add_argument("--layers", type=str, nargs='*', default="-1",
-                        help="Representation layers to extract from the model. Default is the last layer. Example: argument '--layers -1 6' will output the last layer and the sixth layer.")
-    PARSER.add_argument('--pooling', type=lambda x: (str(x).lower() == 'true'), default=True,
-                        help="Whether to pool the embeddings or not. Default is True.")
-    return PARSER.parse_args()
-
-
-
-#debug values
-#os.environ["PYTORCH_CUDA_ALLOC_CONF"]='expandable_segments:True'
-#FASTA_FILE = '/doctorai/userdata/airr_atlas/data/sequences/test_500.fa'
-#OUTPUT_FILE = '/doctorai/userdata/airr_atlas/embedding/test/test_cdr3.pt'
-#LAYERS = list(range(1,33 + 1))
-#LAYERS = list(range(1,33 + 1))
-#CDR3_PATH = '/doctorai/userdata/airr_atlas/data/sequences/trastuzumab/trastuzumab_cdr3_heavy.csv'
-#CONTEXT = None
-#POOLING = False
-
-def fasta_to_dict(fasta_file):
-    """Convert FASTA file into a dictionary."""
-    print('Loading and batching input sequences...')
-    seq_dict = {}
-    with open(fasta_file) as f:
-        for record in SeqIO.parse(f, 'fasta'):
-            seq_dict[record.id] = " ".join(str(record.seq)) # AA tokens for hugging face models must be space gapped
-            # print progress
-            if len(seq_dict) % 1000 == 0:
-                print(f'{len(seq_dict)} sequences loaded')
-    return seq_dict
-
-def load_cdr3(cdr3_path):
-    """Load CDR3 sequences and store in a dictionary."""
-    if cdr3_path:
-        with open(cdr3_path) as f:
-            reader = csv.reader(f)
-            cdr3_dict = {rows[0]:rows[1] for rows in reader}
-        return cdr3_dict
-    else:
-        return None
+BATCH_SIZE = 20000 # works with Nvidia V100-32GB GPU
 
 def initialize_model(model_name=MODEL_NAME):
-    # Loading the pretrained model and alphabet for tokenization
+    """Initialize the model, tokenizer"""
+    #  Loading the pretrained model and alphabet for tokenization
     print("Loading model...")
     model, alphabet = pretrained.load_model_and_alphabet(model_name)
     model.eval()  # Setting the model to evaluation mode
@@ -97,7 +46,6 @@ def initialize_model(model_name=MODEL_NAME):
     if torch.cuda.is_available():
         model = model.cuda()
         print("Transferred model to GPU")
-        device = "cuda"
     return model, alphabet
 
 def load_data(fasta_path, alphabet, batch_size=BATCH_SIZE):
@@ -189,30 +137,6 @@ def compute_embeddings(data_loader, batches, model, layers, sequences, context, 
     torch.cuda.empty_cache()
     return mean_representations, sequence_labels
 
-def export_embeddings(mean_representations, layers, output_path, context=None, pooling=True):
-    """Stack representations of each layer into a single tensor and save to output file."""
-    for layer in layers:
-        if context:
-            output_file_layer = output_path.replace('.pt', f'_context_{context}_layer_{layer}.pt')
-        else:
-            output_file_layer = output_path.replace('.pt', f'_layer_{layer}.pt')
-        if pooling:
-            mean_representations[layer] = torch.vstack(mean_representations[layer])
-        else:
-            output_file_layer = output_file_layer.replace('.pt', '_full.pt')
-        torch.save(mean_representations[layer], output_file_layer)
-        print(f"Saved mean representations for layer {layer} to {output_file_layer}")
-
-def export_sequence_indices(sequence_labels, output_path):
-    """Save sequence indices to a CSV file."""
-    output_file_idx = output_path.replace('.pt', '_idx.csv')
-    with open(output_file_idx, 'w') as f:
-        f.write('index,sequence_id\n')
-        for i, label in enumerate(sequence_labels):
-            f.write(f'{i},{label}\n')
-    print(f"Saved sequence indices to {output_file_idx}")
-
-
 def main():
     args = parse_arguments()
     # Store arguments
@@ -220,8 +144,18 @@ def main():
     output_path = args.output_path
     cdr3_path = args.cdr3_path
     context = args.context
-    layers = list(map(int, args.layers[0].split()))
+    layers = list(map(int, args.layers))
     pooling = bool(args.pooling)
+    # Print summary of arguments
+    print(f"FASTA file: {fasta_path}")
+    print(f"Output file: {output_path}")
+    if cdr3_path:
+        print(f"CDR3 file: {cdr3_path}")
+    if context:
+        print(f"Context: {context}")
+    print(f"Layers: {layers}")
+    print(f"Pooling: {pooling}\n")
+
 
     # Read sequences from the FASTA file
     sequences = fasta_to_dict(fasta_path)
